- Convert the file `DataPreviewMapOpenInNationalMapButton` to TSX file and added TS definitions - Send terria v8 data format by default but v7 format when config.supportExternalTerriaMapV7 = true - Set CI deploy config to send data in v7 format
